Config: Move config mocks in own folder
To clean up the config folder, all mocks are now in it's own folder.
