+++ test/public-api/me.e2e-spec.ts
@@ -30,8 +30,8 @@ import { AuthModule } from '../../src/auth/auth.module';
 import { UsersModule } from '../../src/users/users.module';
 import { HistoryModule } from '../../src/history/history.module';
 import { ConfigModule } from '@nestjs/config';
-import mediaConfigMock from '../../src/config/media.config.mock';
-import appConfigMock from '../../src/config/app.config.mock';
+import mediaConfigMock from '../../src/config/mock/media.config.mock';
+import appConfigMock from '../../src/config/mock/app.config.mock';
 import { User } from '../../src/users/user.entity';
 import { MediaService } from '../../src/media/media.service';
 import { MediaModule } from '../../src/media/media.module';
