+++ test/public-api/media.e2e-spec.ts
@@ -16,8 +16,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { promises as fs } from 'fs';
 import * as request from 'supertest';
 import { PublicApiModule } from '../../src/api/public/public-api.module';
-import mediaConfigMock from '../../src/config/media.config.mock';
-import appConfigMock from '../../src/config/app.config.mock';
+import mediaConfigMock from '../../src/config/mock/media.config.mock';
+import appConfigMock from '../../src/config/mock/app.config.mock';
 import { GroupsModule } from '../../src/groups/groups.module';
 import { LoggerModule } from '../../src/logger/logger.module';
 import { NestConsoleLoggerService } from '../../src/logger/nest-console-logger.service';
